move utility function to infra to fix i386
